or's changes

+++ Ex03.GarageLogic/Wheel.cs
@@ -9,10 +9,10 @@ namespace Ex03.GarageLogic
     public struct Wheel
     {
         internal string m_WheelMaker;
-        internal byte m_CurrentAirPressure;
-        internal byte m_MaxAirPressure;
+        internal float m_CurrentAirPressure;
+        internal float m_MaxAirPressure;
 
-        public Wheel(string i_WheelMaker, byte i_CurrentAirPressure, byte i_MaxAirPressure)
+        public Wheel(string i_WheelMaker, float i_CurrentAirPressure, float i_MaxAirPressure)
         {
 
             m_WheelMaker = i_WheelMaker;
